requirements.txt
asgiref==3.8.1
build==1.2.2.post1
click==8.1.8
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
names==0.3.0
packaging==24.2
pyproject_hooks==1.2.0
pyroute2==0.8.1
python-iptables==1.0.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
unshare==0.22
python-engineio==4.11.2
python-socketio==5.12.1
